haku-console frontend: split client.ts per integration, dedupe ApprovalQueueItem
Per review feedback: client.ts mixed generic capability/approval fetchers
with Gmail-specific (fetchGmailThreadPreviews) and Grocy-specific
(fetchGrocyReference) ones. Split into google_client.ts / grocy_client.ts,
mirroring the existing per-server pattern (google_tool_previews.tsx,
grocy_tool_previews.tsx); client.ts now exports `api` and `errorDetail` as
the shared seam the two new files build on.

Also: ApprovalQueueItem's "tool"/"geolocation" variants duplicated `id` and
`createdAt` instead of sharing them from a base type — same disc-union smell
already fixed on the backend (AliveToolMetadata/DegradedToolMetadata,
AliveServerMetadata/DegradedServerMetadata, grocy_mcp's GetOk/GetError).
