+++ haku/console/frontend/google_tool_previews.tsx
@@ -19,7 +19,7 @@ import {
   zCreateGmailDraftArgs,
   zEventDateTime,
 } from "./api/schema.zod.ts";
-import { fetchGmailThreadPreviews, type GmailThreadPreview } from "./client.ts";
+import { fetchGmailThreadPreviews, type GmailThreadPreview } from "./google_client.ts";
 import { Field } from "./field.tsx";
 
 const GOOGLE_SERVER_ID = "google";
